remove silly class prop consoles and move API files as requested
