+++ zoo_frontend/pages/print/labels/admin.jsx
@@ -18,7 +18,7 @@ class Home extends Component {
     this.state = {
       asdf: props.token, // eslint-disable-line react/no-unused-state
     };
-    console.log(props.classes);
+    
   }
 
   render() {
