+++ zoo_frontend/src/pages/admin/admin.jsx
@@ -23,7 +23,7 @@ class Home extends Component {
     this.state = {
       asdf: props.token, // eslint-disable-line react/no-unused-state
     };
-    console.log(props.classes);
+    
   }
 
   render() {
