+++ zoo_frontend/src/pages/department/edit/edit.jsx
@@ -21,7 +21,7 @@ class Home extends Component {
     this.state = {
       asdf: props.token, // eslint-disable-line react/no-unused-state
     };
-    console.log(props.classes);
+    
   }
 
   render() {
