+++ zoo_frontend/src/pages/food/new/new.jsx
@@ -21,7 +21,7 @@ export default class extends Component {
     this.state = {
       asdf: props.token, // eslint-disable-line react/no-unused-state
     };
-    console.log(props.classes);
+    
   }
 
   render() {
